Continuing updating times from float to double
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@6080 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/columns/HyPerColRunDelegate.hpp
@@ -26,7 +26,7 @@ public:
    HyPerColRunDelegate();
    virtual ~HyPerColRunDelegate();
 
-   virtual void run(float time, float stopTime) = 0;
+   virtual void run(double time, double stopTime) = 0;
 };
 
 }
